last

+++ src/resorces/posts/controller.js
@@ -2,7 +2,7 @@ const Post = require('./model');
 
 exports.getAllPosts = async (req, res, next) => {
     try {
-        const posts = await Post.find();
+        const posts = await Post.find().populate({ path: "user_id", select: "name" });
         res.status(200).json({
             data: posts,
         });
@@ -18,7 +18,8 @@ exports.createNewPost = async (req, res, next) => {
         const newPost = await Post.create({
             title,
             body,
-            tags
+            tags,
+            user_id: req.user_id,
         });
         res.status(200).json({
             message: "Created successfully",
@@ -27,7 +28,7 @@ exports.createNewPost = async (req, res, next) => {
 
     } catch (e) {
         next(e);
-    };
+    }; 
 };
 
 exports.getPostById = async (req, res, next) => {
@@ -46,6 +47,14 @@ exports.getPostById = async (req, res, next) => {
 exports.deletePostById = async (req, res, next) => {
     const id = req.params.id;
     try {
+        const post = await Post.findById(id);
+        if (req.user_id.toString() !== post.user_id.toString()) {
+            res.status(401).json({
+                "message": "Can't do this action"
+            });
+            return;
+
+        };
         await Post.findByIdAndDelete(id);
         res.status(200).json({
             "message": "Post deleted successfully"
@@ -60,6 +69,14 @@ exports.updatePostById = async (req, res, next) => {
     const { title, body, tags } = req.body;
     const id = req.params.id;
     try {
+        const post = await Post.findById(id);
+        if (req.user_id.toString() !== post.user_id.toString()) {
+            res.status(401).json({
+                "message": "Can't do this action"
+            });
+            return;
+
+        };
         const updatePost = await Post.findByIdAndUpdate(id,
             { title, body, tags },
             { new: true }
